Send schedule updates to twitter
We added some more logging earlier so also updated README with current
status around that.

For fixing #10 and adding some #twitter notifications:

Massive library import needed, could not add tweepy as a library: in
app.yaml and we are indeed running this on a static AppEngine:

https://cloud.google.com/appengine/docs/standard/python/tools/using-libraries-python-27#vendoring

No _ssl module found (builtin to python) when running tweepy:

https://stackoverflow.com/questions/5128845/importerror-no-module-named-ssl

Setting environment variables involve deploying and not pushing the
env_variables.yml to github:

https://cloud.google.com/composer/docs/how-to/managing/environment-variables

Also we log that we tweeted and update cloud.md and mention env_variables in case we ever need to bootstrap this from scratch again / or change provider.